Final code clean up
